+++ lib/transpec/record.rb
@@ -1,12 +1,15 @@
 # coding: utf-8
 
+require 'transpec/annotatable'
+
 module Transpec
   class Record
-    attr_reader :original_syntax, :converted_syntax
+    attr_reader :original_syntax, :converted_syntax, :annotation
 
-    def initialize(original_syntax, converted_syntax)
+    def initialize(original_syntax, converted_syntax, annotation = nil)
       @original_syntax = original_syntax
       @converted_syntax = converted_syntax
+      @annotation = annotation
     end
 
     def ==(other)
@@ -23,6 +26,20 @@ module Transpec
 
     def to_s
       "`#{original_syntax}` -> `#{converted_syntax}`"
+    end
+  end
+
+  class Annotation
+    include Annotatable
+  end
+
+  class AccuracyAnnotation < Annotation
+    def initialize(source_range)
+      message = "The `#{source_range.source}` has been converted " \
+                'but it might possibly be incorrect ' \
+                'due to a lack of runtime information. ' \
+                "It's recommended to review the change carefully."
+      super(message, source_range)
     end
   end
 end