+++ lib/transpec/syntax/method_stub.rb
@@ -64,7 +64,7 @@ module Transpec
         return if method_name == :stub_chain && !rspec_version.receive_message_chain_available?
 
         unless allow_to_receive_available?
-          fail ContextError.new(selector_range, "##{method_name}", '#allow')
+          fail ContextError.new("##{method_name}", '#allow', selector_range)
         end
 
         source, type = replacement_source_and_conversion_type(rspec_version)